refactor: edit menu
-Refactor logic for editing sensors and devices and move it into seperate components for better maintainability.
